@@ -1,5 +1,3 @@
-library measured_size;
-
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
@@ -16,9 +14,9 @@ class MeasuredSize extends StatefulWidget {
   final OnWidgetSizeChange onChange;
 
   const MeasuredSize({
-    Key key,
-    @required this.onChange,
-    @required this.child,
+    Key? key,
+    required this.onChange,
+    required this.child,
   }) : super(key: key);
 
   @override
@@ -28,13 +26,13 @@ class MeasuredSize extends StatefulWidget {
 class _MeasuredSizeState extends State<MeasuredSize> {
   @override
   void initState() {
-    SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
+    SchedulerBinding.instance!.addPostFrameCallback(postFrameCallback);
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
+    SchedulerBinding.instance!.addPostFrameCallback(postFrameCallback);
     return Container(
       key: widgetKey,
       child: widget.child,
@@ -45,11 +43,11 @@ class _MeasuredSizeState extends State<MeasuredSize> {
   var oldSize;
 
   void postFrameCallback(_) async {
-    var context = widgetKey.currentContext;
+    var context = widgetKey.currentContext!;
 
     await Future.delayed(
-        Duration(milliseconds: 10)); // wait till the image is drawn
-    Size newSize = context.size;
+        Duration(milliseconds: 100)); // wait till the image is drawn
+    Size newSize = context.size!;
     if (newSize == Size.zero) return;
     if (oldSize == newSize) return;
     oldSize = newSize;
